Move 'src/index.js' to base directory as 'App.js'
also remove .babelrc because I don't think it's necessary
